requirements.txt
numpy~=1.22.3
matplotlib~=3.5.1
pillow~=9.0.1
scipy~=1.8.0
opencv-python~=4.5.3.56
pyqt5~=5.15.4
PyMaxFlow==1.2.13